Pura browser.ts pienempiin palasiin
