+++ valpas-web/test/integrationtests/router.test.ts
@@ -1,7 +1,5 @@
-import {
-  goToLocation,
-  textEventuallyEquals,
-} from "../integrationtests-env/browser"
+import { textEventuallyEquals } from "../integrationtests-env/browser/content"
+import { goToLocation } from "../integrationtests-env/browser/core"
 
 describe("Reititys", () => {
   it("Näyttää virhesivun kun reittiä ei löydy", async () => {
